tests/smtpd.py: Returns a permanent failure for unsupported auth mechanisms
smtpd testing server does call a specific method when AUTH PLAIN is
used, but ignores without returning any other auth mechanism. As the
connector tries auth methods in a non-deterministic order, this leads to
a timeout if PLAIN is not the first mechanism tried.

This commit makes the smtpd test server return a permanent failure for
non-supported auth mechanisms.

Fixes issue #171